scorer/requirements.txt
# Core API framework
fastapi
uvicorn[standard]
python-multipart

# Audio processing & ML
torch
torchcodec
torchaudio
py-espeak-ng
soundfile

# Basic utilities
numpy